+++ apps/web/src/components/Pages/Contact.tsx
@@ -1,7 +1,6 @@
 import MetaTags from '@components/Common/MetaTags';
 import SettingsHelper from '@components/Shared/SettingsHelper';
 import { EmptyState } from '@components/UI/EmptyState';
-import { Form, useZodForm } from '@components/UI/Form';
 import { GridItemEight, GridItemFour, GridLayout } from '@components/UI/GridLayout';
 import { Input } from '@components/UI/Input';
 import { Spinner } from '@components/UI/Spinner';
@@ -15,7 +14,7 @@ import { APP_NAME, FRESHDESK_WORKER_URL } from 'data/constants';
 import type { FC } from 'react';
 import { useEffect, useState } from 'react';
 import { PAGEVIEW } from 'src/tracking';
-import { Button, Card } from 'ui';
+import { Button, Card, Form, useZodForm } from 'ui';
 import { object, string } from 'zod';
 
 const newContactSchema = object({
