+++ apps/web/src/components/Settings/Account/SuperFollow.tsx
@@ -1,4 +1,3 @@
-import { Form, useZodForm } from '@components/UI/Form';
 import { Input } from '@components/UI/Input';
 import { Spinner } from '@components/UI/Spinner';
 import { StarIcon, XIcon } from '@heroicons/react/outline';
@@ -19,7 +18,7 @@ import { useState } from 'react';
 import toast from 'react-hot-toast';
 import { useAppStore } from 'src/store/app';
 import { SETTINGS } from 'src/tracking';
-import { Button, Card } from 'ui';
+import { Button, Card, Form, useZodForm } from 'ui';
 import getSignature from 'utils/getSignature';
 import getTokenImage from 'utils/getTokenImage';
 import { useContractWrite, useSignTypedData } from 'wagmi';
