+++ apps/web/src/components/Shared/Login/New/index.tsx
@@ -1,6 +1,5 @@
 import ChooseFile from '@components/Shared/ChooseFile';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
-import { Form, useZodForm } from '@components/UI/Form';
 import { Input } from '@components/UI/Input';
 import { Spinner } from '@components/UI/Spinner';
 import { PlusIcon } from '@heroicons/react/outline';
@@ -10,7 +9,7 @@ import { APP_NAME, HANDLE_REGEX, ZERO_ADDRESS } from 'data/constants';
 import { useCreateProfileMutation } from 'lens';
 import type { ChangeEvent, FC } from 'react';
 import { useState } from 'react';
-import { Button } from 'ui';
+import { Button, Form, useZodForm } from 'ui';
 import getStampFyiURL from 'utils/getStampFyiURL';
 import { useAccount } from 'wagmi';
 import { object, string } from 'zod';
