+++ apps/web/src/components/Shared/Modal/Report/index.tsx
@@ -1,6 +1,5 @@
 import { EmptyState } from '@components/UI/EmptyState';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
-import { Form, useZodForm } from '@components/UI/Form';
 import { Spinner } from '@components/UI/Spinner';
 import { TextArea } from '@components/UI/TextArea';
 import { PencilAltIcon } from '@heroicons/react/outline';
@@ -14,7 +13,7 @@ import type { FC } from 'react';
 import { useEffect, useState } from 'react';
 import { useGlobalModalStateStore } from 'src/store/modals';
 import { PAGEVIEW, PUBLICATION } from 'src/tracking';
-import { Button } from 'ui';
+import { Button, Form, useZodForm } from 'ui';
 import { object, string } from 'zod';
 
 import Reason from './Reason';
